Update to using zuulv3 variables
The prepare-docs-for-afs role was using zuulv2 variables,
but these don't exist in zuulv3. Update to using the variables
that do exist. Add a flag to take the place of $2 in the original
script which was called by the docs-tags-only-env builder.

When we migrate, we can just add that variable to the
openstack-doc-build entry in the pipelines of bindep, git-restack and
git-review which are the only three projects which use that flag.

Change-Id: Ib998f83135a4fcb796a205e52c435ab29b7cab7b
Signed-off-by: Paul Belanger <pabelanger@redhat.com>