Create legacy nodesets and switch all legacy jobs
The legacy jobs were designed to pull logs only from their 'primary'
node. The converted jobs are pulling them from all of the nodes.

Create legacy- versions of all the nodeset that name the single node
"primary". Update all of the legacy jobs to use these. The next patch
will update all of the post playbooks to be hosts: primary.

Legacy versions of multi-node nodesets aren't strictly needed as the
non-legacy versions already have a 'primary' ... but it makes both
global search and replace as well as migration docs simpler.

Change-Id: Ic04f38d9b5251f5a1a5426f890728bf0dd9b4e9c