Fix tripleo multinode job defs
This may not fix all the problems in the tripleo jobs but adds in the
base set of things we've learned about making multinode jobs work in the
migrated zuulv3 legacy jobs.

First we only run the main run playbook against the primary node.
Otherwise it runs on all the nodes and gets confused.

Second we open up the firewall between all the nodes so that they can
talk to each other.

Change-Id: Ibcba1bd338eeb7e9ccf983085838ef2a65ab647f