Cleanup kolla-ansible multinode jobs
Don't know if this is sufficient to get kolla ansible multinode jobs
working but it does fix a couple items that we have discovered with the
migrated multinode jobs.

First use a base job that opens up the firewall between all tests nodes.

Second only run the main run playbook against the primary test node.
This is how it worked in v2 so running the playbook against all nodes
creates problems.

Change-Id: Iffc52b17fb0e88985364d63cd32c93c263914074