Move tox_envlist variable for releasenotes into job
The tox_envlist for the job that builds release note was
set in the post.yaml to fetch the tox output only.  However,
becasue it was not specified in the job, tox-siblings which
runs in the `pre` stage of `tox` would attempt to use `venv`
in the envlist field.

For jobs that do not have that venv (but do use release notes),
it would result in a failure.  This patch moves the variable out
into the job so that it's applied for the `tox-siblings` role.

Change-Id: If660822c763a973963be0c283dfdbebd97f06985