Avoid moving release notes after complete execution
Originally, we were moving the release notes to another folder
to make it work with the fetch-sphinx-output role.  However,
this did not always work as expected due to the fact that some
projects do not have a doc folder.

This fixes that issue and avoids the move and simply grabs the
data from the right folder.

Change-Id: If6ee519c33320db91f8a9553ed4abc099c284eec