Stop doing afs prep in docs build jobs
We don't publish docs to AFS in the build jobs, so the step is
unnecessary and also prevents infra projects from using the normal docs
build job.

Doing this means we're not reorganizing the docs, so we can remove
latest/ from the success-url - and we can also remove the
branch-specific variants.

Remove mention of docs build jobs from the
publish-openstack-sphinx-docs-infra project-template. The build jobs are
in openstack-python-jobs.

A followup will come to change that - but for now let's align these.

Change-Id: I123b4d9f3514542d8ece8e9681c5265dc54961dc