Put doc build jobs into doc publication project-template
They aren't python specific, and it's important to run them in check and
gate if one is intending to run them in post/release.

Add a template for neutron, because of the neutron needing neutron
issue.

Add a template for just building and not publishing.

Update the template for docs-on-readthedocs for projects who want to
follow the docs PTI but aren't official and and so publish to readthedocs.

Change-Id: I8f6a0b1401099def47bb349c5c2453429bd19085