Don't remove $WORKSPACE on propsal jobs
These jobs used to run on a long-lived node, and we needed to periodically
clean out the jenkins home dir of workspaces or the disk would fill.  Now
they run on single-use nodes and we don't.  Moreover, removing everything
in the homedir also removes the prepared git repos in v3, which makes the
jobs fail.

Change-Id: I86bc40885abea73bb90aee3998c1938e26f6e1f8