Add zuul-jobs to required-projects for tox-linters
The tox-linters job for openstack-zuul-jobs pulls in the contents of
zuul-jobs so that it can validate against roles that come from
zuul-jobs. In cases where an openstack-zuul-jobs patch depends on a new
role in zuul-jobs, the depends-on doesn't work properly and the linters
job in openstack-zuul-jobs will fail due to missing roles from
zuul-jobs.

tox-siblings should be able to take care of this for us, so just add
zuul-jobs to required-projects.

Change-Id: Ie55689ede6763e59a0dfeca758632818ac45fd1b