Add branches to all periodic jobs
Periodic pipelines emit events for all branches. These jobs all only
need to run on their respective branches. Add branches: matches to limit
them to only running on those branches.

When we replace these with new jobs, we should also collapse, so that we
just use 'build-openstack-sphinx-docs' and put it into the periodic
pipeline - and limit it to the branches necessary when we put it into
the pipeline.

Change-Id: Iae3570d8dc69dda9be11df4ff4cd89003b971896