Move tox_envlist into job variables for releasenote jobs
At the moment, the tox_envlist is defined inside the include_role
for tox as releasenotes.  This means that other roles which run
beforehand such as tox-siblings will use the default value of
venv instead which isn't always defined.

This patch makes the change to make that variable defined in
the job which means that all other roles will reference the correct
tox envlist in the entire job run.

Change-Id: I0323f8f1bccf720268d947b8e1dcb6c5f8b26dfa