Ensure func->functional tox env mapping works
Currently the mapping of the 'functional' to the 'func'
tox environments does not work for Ubuntu Xenial. This
is due to the shell not being set, and the default shell
somehow being 'dash' instead of 'bash'.

This ensures that the shell is properly set for all shell
tasks, but also makes this mapping optional. If the
'functional' tox env is present it will be used - otherwise
it will use the 'func' tox env. This allows all these jobs
to transition if they wish.

Change-Id: Id63e626fd34f9c2ebb89dfae8fdda936a1ecfdf5