Refactor fact configuration for service_type_data
The service_type_data was not working properly and errored out
with a traceback on exit_json from the module.

We can actually handle this in native Ansible without a module so
let's do that and also hint the user at the issue if a project happens
to use this role without being mapped in the service types.

Change-Id: Icdac2f5325a01748cdb8a830430b8f66a035d416