Add configure-swap role
This is a role to configure swap and /opt on an infra host.

Originally we wrote this for devstack-gate with
Icd4e6d8ab84471ad06e69c3e0f9bac92776efc78.  It was ported into the
native devstack job with Iffe54fbccbccd68db08f79a1b51dd7f76dbff408.

This is really generic, and it's quite possible non-devstack jobs
would like to setup swap for jobs too, so o-z-j is a better home for
the role.

Change-Id: I0e9c846ace7fac8a1340746c6818fba6ec963018