Add missing required-projects for puppet-openstackci
This unifies the beaker and openstackci puppet rspec/beaker
jobs under a same base job for centralizing the required projects.

This resolves failing gate jobs for puppet-openstackci since it
uses system-config to install modules which, in turn, uses zuul-cloner.

We're also turning "legacy-openstackci-beaker" back to voting because
it has been shown to work properly with this set of required-projects.

Change-Id: I9d0b2ffb1734da11c7e362c10bd2e4c643805d13