Run gerrit periodic on openstack/2.13 only
We have several openstack/2.13 branches, the v2 job run only on exactly
one - the openstack/2.13 one. Let's run the new ones on exactly the same
branch.

Note that branches is a regex, see
https://docs.openstack.org/infra/zuul/feature/zuulv3/user/config.html#attr-job.branches
So, we need to add ^ and $ to be explicit and avoid match against
openstrack/2.13.3 and similar named branches.

Change-Id: I88fb6d5e7df1739724901b661a9365c30582913c