Create dir needed to publish gerrit plugins
rsync will only create a single directory to copy contents in so we need
to create every dir level above that. In this case are copying to
$workdir/artifacts/plugins/$plugin and were only creating the artifacts
dir and rsync will only attempt to mkdir $plugin so this fails on
missing the plugins dir.

Fix this by also creating the plugins dir.

Change-Id: I93ecb3d39cc751bb7d4a17173fd02e41ea77b620