Use nodejs-npm jobs
Use new native nodejs jobs for lint and test in
project-templates. Move nodejs native v3 templates to project-templates
file.

There are some additional project specific legacy nodejs jobs that
install chromium and xvfb in addition, converting them is outside of the
scope of this change.

Note that the legacy template run node6 only in check as voting job, add
it to gate as well now.

Add a TODO to the nodejs docs job templates, they are wrong right now.

Depends-On: I45294049d9228020b70f6b18265e21fa9385f782
Change-Id: Ib38294ab48ef158a248637de79970ef60bc44f7f