Switch build-openstack-sphinx-docs to build-sphinx-docs
As per the recent PTI update, docs should be built without tox and
instead using direct sphinx-build commands. The general job
build-sphinx-docs does this, so use it. Keep the
build-openstack-sphinx-docs job because we need to set constraints
support.

This includes a hack to allow neutron and horizon plugin repos to work
properly with their tox_install.sh list of additional requirements.

Change-Id: Ib121961c5a953a434e7b333cd70f7838a2671f69
Depends-On: Ie65dcb42c48e6a962f6715f7483ef3758caf2965