Remove required-jobs from final docs job
We cannot modify the final job, remove the required-jobs. This will
break the publishing - but right now *no* post job runs for these and
that's even worse.

We would need to remove the required-project lines from check and gate
as well but if we do this, they fail completely and thus repos using
this cannot merge anything. Let's do this later once the jobs are
changed to not use tox - see topic:updated-pti.

Change-Id: I5bc1ac2816c1a7bba425d06db9e8bb8bdf025c32