Add support for configuring TTLs in configure-unbound
Given the volume and ephemeral nature of the nodes we are running jobs
on, we're interested in raising the minimum TTL to ensure we're not
needlessly querying domains all the time.

Some domains such as github.com or fedoraproject.org have low TTLs and
we suspect there is a correlation between their low TTLs and the fact
that we're seeing increased DNS lookup rates for these domains.

Change-Id: I6d9656cb07a694fc2f54f256a63af814c034ffb8