Add required projects to legacy puppet rspec jobs
These rspec jobs install the entire set of modules declared by
system-config's modules env file. It also does so from source using
zuul-cloner so that we can do cross repo testing of our puppet modules.
Unfortunately these required projects were not listed on the jobs
resulting in errors from the zuul cloner shim when attempting to clone
these repos.

Add in all our puppet modules as required projects (as listed by the
zuul cloner shim) so that we fix this problem. We do this via a new job
that is infra specific so that we don't have to modify the required
projects list in multiple places to get the correct list.

Change-Id: Iff50f517bb92fd19a620ab5b433cd2e52702cec0