Make infra specific beaker rspec template
Turns out the puppet openstack modules do run the legacy xenial beaker
rspec job. I had thought they were completely off of the legacy job
config but was wrong. This means we need new templates for infra that
add in the infra repos and need to restore the existing templates that
don't add in the infra repos for the puppet openstack modules.

Change-Id: Icf7ec289d51a3e7b7374f8aef5a94d18ec12bc94