Skip jobs for driverfixes branches
There were a few jobs added to driverfixes runs with the conversion
to zuulv3 that are not needed/wanted. The plan for these was to
limit to just pep8 and unit test checks.

Since exclusions can't be done on the project specific variants, this
adds the exclusions to the job definitions themselves. This should be
OK since currently the only users of driverfixes branches have the
same policy, and most of these do not make sense to run against
something like these branches.

One thing to note is I am limiting unit test runs to just the py27
job. We may want to think about changing this in the future, but
for now we only need one version run and, at least for these older
versions, py27 is the more prevalent environment.

Change-Id: I21b3a0b924709b319c0aa67347242aa806a47117