Add heat-tempest-plugin to grenade/sahara job
Otherwise the heat grenade plugin, which is enabled and used by Sahara,
fails, because the heat tests have been split but they are used by the
grenade plugin.

It is true that this could be migrated locally to sahara repositories,
but the effort is to try to push native jobs directly.

Change-Id: I15d1c25f5701e8fd3450f61cc661515639ba434a