base/multinode rename #1: Add renamed jobs
We want to rename the base and multinode integration jobs in order
to make sure they are not mistaken by users trying to test their
projects. These jobs are not meant to be used outside of integration
testing the playbooks and roles found in project-config, zuul-jobs and
openstack-zuul-jobs.

This is part one of three in renaming the base and multinode
integration jobs. We need to:
1) Add new jobs in openstack-zuul-jobs        <-- We're here
2) Make project-config use the new job names
3) Remove old jobs

Change-Id: I150a0a6dacc7b8862a40f1382ea730957dea0faf