base/multinode rename #3: Remove old jobs
We want to rename the base and multinode integration jobs in order
to make sure they are not mistaken by users trying to test their
projects. These jobs are not meant to be used outside of integration
testing the playbooks and roles found in project-config, zuul-jobs and
openstack-zuul-jobs.

This is part three of three in renaming the base and multinode
integration jobs. We need to:
1) Add new jobs in openstack-zuul-jobs
2) Make project-config use the new job names
3) Remove old jobs                            <-- We're here

Depends-On: I4ef44e64a03cc3089e02343de506e0a6fd85a55c
Change-Id: Ic251346f3a2ec51e49ca07417e9ba84ec1d3a4af