puppet-beaker tests: don't use puppet 4
The original job code was doing a "if {node} == ubuntu-xenial".  This
got expanded out by the autoconversion, but because it uses "==" with
"[" and runs under POSIX shell (rather than bash) this just ends up as
a syntax error.

Since the test is a no-op in most cases, remove it to avoid the error.
Also use /bin/bash for the cmd for sanity.

This also makes the Xenial tests just use puppet3, which is the puppet
included with Xenial.  This is becoming an issue because infra is
updating to Xenial nodes now, so we are using this against some of our
puppet modules, but only targetting puppet3.

Change-Id: I2355cf58c899bf5f78173b2ed9da26548a9592d0