Rework translation templates
Currently we run the post job for each and every change, even if that
branch is not set up in Zanata for translation.

Since we have already branch specific templates, let's use them and add
a post job for each of them that mirrors the periodic job.

This way we only run a post job on a branch if we also have a
corresponding periodic job to import changes back.

Change-Id: I6b647ecee41f1d41f0d59f913d5213046b253b22