Add horizon,neutron to build-openstack-sphinx-docs
The neutron-horizon hack is run on the base build-openstack-sphinx-docs
job. Add the required repositories neutron and horizon to it.

We could then remove the repositories from its usage in the
project-templates but will not do so. Change
I57ace5de5fd36efe2e513674734e97a0048743b6 will remove the usage of
those templates and then we can delete them.

Change-Id: Ia09e0d57752aeafc322857e9a98a3e1dba77b745