Make sure we only run tempest-full on master
Since tempest-full is part of the integrated gate, it is executed
against all stable branches. However since stable branches will
not get roles from master anymore this does not work.

We may want to backport all devstack ansbile changes, but in the
meanwhile this ensures that a working job is executed against
stable branches.

Change-Id: I5d34e5b983b720dc1913d00ccc1a16d81a040d49
Partial-bug: #1739829