Create artifacts dir regardless of afs rpm builds
We don't always create afs rpms in our periodic job so we were skipping
the creation of the artifacts dir. Unfortunately the openstack
artifact publishing playbook expects this dir to exist. Fix this by
creating the artifacts dir but not populating it with RPMs if we didn't
build any RPMs for openafs.

Change-Id: I986bca765ae898e9792e789c6852bce1d654d580