Remove build-javascript-tarball job from nodejs-jobs
The sole purpose of the build-javascript-tarball job is to check that
"npm pack" succeeds.  In an npm project, "npm pack" depends on "npm
install".  Once the project is successfully installed, "npm pack" always
succeeds because it's just a thin wrapper around "tar".
build-javascript-tarball has nodejs-npm as its parent which handles
installing of npm dependencies.

Therefore, build-javascript-tarball is redundant.

Change-Id: Id57e4c41f23544891f2a16cba3f4b5c59b4610c1