Don't run legacy-tempest-dsvm-cells on newton/ocata/pike
For some reason, the cellsv1 job has been failing at a very
high rate within the last couple of weeks but only on stable
branches; libvirt connections have been dropped causing test
failures. This isn't a problem on master, for whatever reason -
maybe something different in pike UCA packages used on master
for queens, it's not known right now.

We also have changes that we have to merge into stable related
to zuulv3 job definitions to unblock some things on master, so
for now we're going to not run the cellsv1 job on stable branches
until the bug is sorted out, or we redefine the job and make it
non-voting on stable.

Change-Id: Iaed05d569199375ab2db874951aa17b435704295
Related-Bug: #1745838