Replace legacy tempest-dsvm-py35
We now have a tempest-full-py3 zuulv3 native job. Drop the legacy
job in the integrated-gate-py35 and replace with the new native
job. The native job only works on master, and it will work on
stable/queens once the branch is created, but that's ok since
the legacy job only works starting on queens as well.

The branch filters out all stable branches. Updating the brach
logic to exclude all stable branches before stable/queens.

The legacy job is not used anywhere else, so removing it.
The template now uses only new style jobs, so moving it out of
zuul-legacy into project-templates.yaml.

Depends-on: I4a8fa1759665038788e1b86fd3a5030fd6c68573
Change-Id: I785fc270496fbcdd2f18b96f11d0acfa9a83c6ea