Handle tags in infra docs publishing
We publish docs for infra projects on tags for projects like bindep.
Previously the afs publishing role for this assumed zuul.branch was
always going to be present but it isn't when operating on a tag in the
release pipeline. Update the conditionals to properly handle the case of
zuul.tag being present and zuul.branch missing.

Change-Id: Iee2a7f43a92a63c1943755ff46a03c94f096d6bc