Temporarily make legacy-tempest-dsvm-neutron-src NV
Stable/pike is currently broken due to a change in pip [0] and
a change in setuptools [1]. Each change is needed to land to fix
things, but they are in different repos, so there is no way to
combine the changes into one to get it through.

This patch makes the failing job for one non-voting to get that
in, to allow getting in the other one, at which point this one
will be reverted.

[0] https://review.openstack.org/#/c/554270/
[1] https://review.openstack.org/#/c/554053/

Change-Id: I9c46fb9af99bc1960cee513e0cf383f751b0126b