add openstack-tox-lower-constraints
Add a job for running tests for projects using their lower-constraints
list.

These settings are currently based on the ones from the py35 job
because we expect the tox environment to be configured to use python3
wherever that is possible.

We don't set this up as part of the check-requirements job template in
project-config because projects opt-in to running the tests.

http://lists.openstack.org/pipermail/openstack-dev/2018-March/128352.html

Change-Id: I8ad36dfd2ce6d0ab18340c58bd78d234c0d568e6
Signed-off-by: Doug Hellmann <doug@doughellmann.com>