Fix ansible-lint job
The ansible-linter was only running on playbooks, not on roles. This
patch corrects the tox environment to ensure that roles are linted.

Since roles weren't being linted, some errors had sneaked through. This
patch also corrects an unnamed task and disables linting of a task that
deliberately uses the shell module.

Also fix a comment in the tox file that was referencing a non-existent
zuul job.

Change-Id: I5b1ca3947e80588ab0cc8d3632f1292fc707aac1