update the announce-release job to look in releases repo for tools
We are consolidating the release tools into the releases repo to
simplify things and make it possible to test some of the tools against
proposed releases. This change updates the playbook to use the version
of the announce.sh script that is moving into openstack/releases and
to look at the bindep.txt file already in the releases repo.

Change-Id: Ibbee039fd1b784d2c4135120793378acb3193814
Story: #2001816
Task: #14442
Depends-On: https://review.openstack.org/564327
Signed-off-by: Doug Hellmann <doug@doughellmann.com>