More fix laravel branch and master publish jobs
The tarballs are now building and being copied to the executor
correctly, but need to be published to the tarballs site. Reparent
the legacy-laravel-openstackid-release-branch and
legacy-laravel-openstackid-release-master jobs to
legacy-publish-openstack-artifacts instead of legacy-base so the
additional missing publication tasks will be added. Also update the
artifact transit directory name on the executor to match what the
uploader role will expect.

Change-Id: I941b986ee7bf200e2d5a810ab8cefc62567a5b6a