Further fix laravel branch/master publish jobs
The publish-artifacts-to-fileserver role includes the tree structure
of under the artifacts dir, which we don't wish to preserve for the
tarballs directory our legacy-laravel-openstackid-release-branch and
legacy-laravel-openstackid-release-master jobs create. Instead just
copy the contents of the tarballs directory to the artifacts
directory (this also allows us to drop fancy filename pattern
matching and symlink preservation from our rsync options).

Change-Id: I22743c78e286c51067d38efe8c646021e68ca09b