Add Zuulv3 jobs for puppet functional testing
We need to abandon rspec as a functional testing framework. It
constantly breaks as ruby gems change their dependencies, no one likes
to deal with it and it did not help bring new people from the puppet
community to help maintain the modules. This patch adds a Zuulv3-native
job that can act as a framework for the future of puppet functional
testing. Puppet modules will need to have the files
tests/scenarios/*/{manifest.pp,tests.sh} in order to use this.
Eventually this should replace the legacy-puppet-beaker jobs.

Change-Id: Ic47e93fb091a9cdc876981674cff6b452a1afe87