Convert infra's beaker jobs for zuulv3
This patch moves the legacy beaker jobs that were being used for Infra's
puppet modules out of the legecy definitions and into the regular list.

Historically, the test scripts were shared between the Infra and Puppet
OpenStack teams, but the Puppet OpenStack team has already moved their
jobs in-repo so we can remove tasks specific to their setup.

We need to leave some jobs and playbooks in place that are still being
used by the puppet-midonet project which belongs to neither team.

Change-Id: Iefea6ee173cff2bf050d99d96ae2e1d1a51b7f49