Convert puppet-openstackci jobs to zuulv3
This patch moves the puppet-openstackci beaker jobs out of legacy. Since
this job is essentially the same as the regular beaker jobs, we don't
need new playbooks, we just need to adjust the beaker-rspec playbooks to
run the puppet-openstackci tests.

Since this was the last legacy job in the infra-puppet-apply-jobs
template, move it out of the legacy templates to the regular templates.

This doesn't remove the legacy-puppet-openstackci-beaker-rspec jobs
or their parent legacy-openstackci-base since those are still defined in
project-config.

Change-Id: I7f7cb133fc41eba07da224291b7f2d0659b0deb6