Improve kata-runsh job
Add system package dependencies (build-essential, golang-1.10) and
set GOPATH to $HOME for the zuul user which should magically work with
the zuul git repo pathing. Additionally use shell with set -x.

We switch to xenial so that we can use upstream docker-ce packages which
the kata ci seems to expect.

Depends-On: Ief0f20c2f1b42a5c23ed05f7185c58700f65b06e
Change-Id: Ibb5cee5ba90662e53c39dcf686a52bcafe81c933