define publish-openstack-docs-pti project template
As part of the python3-first goal, we need to update the PTI to
restore the ability to use tox to run the documentation build
commands. This will give us a way to configure each repo to have its
dependencies installed properly and to change the version of python
used in a self-testing patch to the tox.ini file.

Depends-On: https://review.openstack.org/580484
Change-Id: If87a3c80bed8dbbe9a93a6350056338e97d98162
Signed-off-by: Doug Hellmann <doug@doughellmann.com>