Use webhook update for docs-on-readthedocs
Update the docs-on-readthedocs template to use the new
trigger-readthedocs-webhook job.

This new job triggers readthedocs via their authenticated webhook API.
It is not a drop-in replacement; users will have to update with a API
key per the documentation (this will be communicated before merge;
however, note that the old API is not working).

For more details, see the comments in the dependent project-config
change.

The ability to have variables in templates is also new, see the
dependent zuul change.

Change-Id: I3b65813671961d07c0a743685c537ad1df4bf682
Depends-On: https://review.openstack.org/583449
Depends-On: https://review.openstack.org/584230