Fix publish-openstack-docs-pti check/gate html
The new publish-openstack-docs-pti template uses the openstack-tox-docs
job which does not save the rendered docs, nor does it set the job link
to the rendered documents.
This patch changes the openstack-tox-docs job to save the rendered html
and set the success-url to the rendered html documents.
This returns the behavior the publish-openstack-sphinx-docs job provided.

Change-Id: I460c00542426f6b24a7423098be9a0af14a35f01