Add warning job about readthedocs publishing
The readthedocs transition is slightly painful because it requires
readthedocs project owners to add infra to their projects, so we can
not do the update on their behalf.  Not everyone reads the mailing
list, and because the jobs run in the post pipeline it is very easily
missed that the publishing isn't actually happening.

Add a fake, non-voting, always failing job that should attract
attention.  Make a new template which we can switch failing projects
to; which they can easily switch back from in the same change that
adds the extra info required to make the jobs work.

Change-Id: I1866b8b00891caa0e6d9cde99555be8b55792e38