Add openstack-tox-functional-py36 job
It's common for projects to now want to run their functional tests
with python 3.6. This patch adds a centralized job for the
functional-py36 tox target, duplicating the functional-py35 job
and adding py36 details based on the py36 job for unit tests. There's
some YAML fanciness that could be done here to avoid duplication
but that doesn't appear to be the fashion, so that's not being
done.

Needed-By: https://review.openstack.org/599558
Change-Id: I02a5ad943ab1241d63fd19205dab9f952ee42d46