Add openstack/placement to legacy-dsvm-base required-projects
Devstack will need the openstack/placement repo starting in Stein
so this adds it to the legacy-dsvm-base required-projects list.
The repo doesn't exist on stable branches before Stein so if zuulv3
checks it out in stable branch job runs it should just be ignored.

Needed by https://review.openstack.org/600162/

Change-Id: Ieeb586f538ecd9d4cbbaabd605825e4c5f06ff63