Update lib-forward-testing
The tempest-full job can replace the legacy job
legacy-tempest-dsvm-neutron-src since tempest installs the checked out
libraries, including the library under test. We have already converted
the py3 variant, and can convert the py2 the same way.

The Zuul v3 native tempest-full job is not available on stable/ocata, so
use the legacy job on ocata and older.

Move the template to project-templates file and add description.

Change-Id: Iac9162c1321f7f825684e1c20cc13ff5821d37d3