Add zuul to test-requirements for linting
The linters env is picking up zuul-jobs content too, so we need to add
zuul to get the zuul_ modules ... otherwise ansible-lint has a sad.

Add zuul to test-requirements like is done in zuul-jobs. Then point the
library path to zuul/ansible/library in the site-packages dir from tox
so that we pick up the right files.

Change-Id: I7a4f0159290d67354e0a73365091a0d1718762fd