Handle operations-guide and contributor-guide
Documentation projects are whitelisted. Since operations-guide and
contributor-guide has a normal "doc/source" setup, we can use the
generic handling of this and just extend the whitelist.
Instead of having the whiteliste in two places, move it in a central
shared place.

This will handle operations-guide and contributor-guide the same way
as horizon, openstack-ansible, and openstack-helm.

Change-Id: Ie1065b7d4ad47dee738568c70df28a5cdaa6f6f7