stop publishing release notes using python 2
Remove the tag pipeline entry that calls
publish-openstack-releasenotes from the release-notes-job project
template so that projects that still have the template in their stable
branches do not trigger multiple jobs to publish release notes when a
release is tagged. We only ever want the notes to be published from
master, and that must stupport publish-openstack-releasenotes-python3
now, so we should not miss any notes.

Change-Id: I1f910b1543f930e8a837a764d5df4f3e95197df0
Signed-off-by: Doug Hellmann <doug@doughellmann.com>