Fix name of 'integrated-gate-py35' project template
Project template 'integrated-gate-py35' was added when OpenStack gate
used to run on xenial distro where python 3.5 is available.

Now we have migrated the gate to Bionic where python 3.6 is available.
With that migration, gate jobs in 'integrated-gate-py35' template are
running on python 3.6 not on 3.5 which makes this template name confusing.

This commit rename 'integrated-gate-py35' to 'integrated-gate-py3'
so that it can convey that template will use available python 3 version
present in used distro.

Renaming the template cannot be done directly as old template name is used in many
project repo and their stable branches. We need to provide new template
'integrated-gate-py3' and once all project will switch to this new
template then, we can remove the old 'integrated-gate-py35' template.

Change-Id: I050a0f81d6f5c00fffa05954ee740840c2a62eff