Add Python3 project templates for Train release
Following the process set down in the TC resolution on updating Python
versions:

https://governance.openstack.org/tc/resolutions/20181024-python-update-process.html

this change adds the Zuul template for Python 3 unit test jobs in the
Train release. The TC has determined that:

* The latest released version of Python that is available for testing is
  3.7.
* The only version of Python 3 that is the default in the
  distributions listed in the PTI (CentOS 7, Ubuntu 18.04 Bionic, openSUSE
  Leap 15) is Python 3.6. (Python 2.7 is the default in CentOS 7.)
* All Xenial-based integration tests will be migrated to Bionic before
  the Train cycle begins, so there is no need to test Python 3.5.

Therefore the Python 3 versions unit tested in Train will be Python 3.6
and Python 3.7. (Python 2.7 will also be tested.)

http://eavesdrop.openstack.org/meetings/tc_python3/2019/tc_python3.2019-03-07-21.00.html

Change-Id: I8b5fc61a6e1c192efed22bfd02155e22187464c1