Use xenial for openstack-tox-docs for all older stable branch
Currently openstack-tox-docs job use ubuntu-xenial nodeset for
stable/rocky and bionic for rest of the branches. But it seems
this job can be run on other stable/branch previous to stable/rocky
also. It is better to use xenial for all the stable branch till rocky.

openstack-tox-docs job is supposed to be run from stable/rocky on
(and build-openstack-sphinx-docs for odler branch) but this
job is also part of docs-on-readthedocs project-template so we need to
adjust the nodeset for older stable branch also.

Change-Id: Iac8a0deea04c60d0d427f7c19a29d9f165d80f26