Update puppet jobs to default to puppet 4
This change removes the last of the puppet-3 puppet-apply jobs and
changes the default puppet version to '4' for the puppet-apply and
beaker jobs. This change leaves the unversioned base puppet jobs and the
puppet-4 child jobs even though they are now redundant for the sake of
clarity of the job ancestry.

Change-Id: I58a82d7e58ee1d2d8c653e4e92cab8cfc8d834b4