Use promote job for infra
For using the promote jobs for publishing docs, change the docs job to
opendev-tox-docs so that the tarball gets generated that is needed for
publishing.

Update  publish-tox-docs-infra to use the promote job, update
release-tox-docs-infra to have the same jobs but no change of publish
job (promote job does not work for release pipeline).

Change-Id: Idc5e94a581359bcfd3a570e085aa13999c26eb86