Add support for building PDFs
We want PDFs. We don't want to have to go to each project and add a
'pdf' tox target or similar. Get the former and avoid the latter by
adding a step to build our PDFs to the 'openstack-tox-docs' job.

The job assumes a 'pdf-docs' tox env to build PDF docs.
If 'pdf-docs' tox env does not exist, PDF doc build will be skipped.

Note that only the job for master and stein is updated,
we do not build PDFs for older releases.

Do not build PDFs for deploy-guide for now.

Co-Authored-By: Stephen Finucane <stephenfin@redhat.com>
Co-Authored-By: Andreas Jaeger <aj@suse.com>
Depends-On: https://review.opendev.org/678077
Change-Id: Ibc455ba511a877159848f4060a6058800d7d2f9b