Use on-disk horizon dir for translations
Instead of checking out horizon ourselves, let Zuul do it for us. The
updated playbooks pass in the horizon directory as parameter.

This fixes manila-ui/train translations since we do not have a
stable/train branch for horizon and therefore the git commit fails. If
we use Zuul for it, it will checkout master as fallback for us.

Depends-On: https://review.opendev.org/682808
Change-Id: Ifa96087916d21be9b4a6fcce99088c5fc1565b41