Explicit mention py2 for grenade py2 jobs
As part of community goal of dropping py27 support[1], we are
moving the devstack to py3 by default[2]. That will make grenade job
to perform upgrade from py2 to py3 which will not work (we have seen
the failure in neutron-grenade job).

These jobs are running on stable branches also which is supposed to run
on py2.

Next step will be to move these jobs on projects side with py3 version
and keep these py2 version also for stable branches.

[1] https://governance.openstack.org/tc/goals/selected/ussuri/drop-py27.html
[2] https://review.opendev.org/#/c/649097/12
    http://lists.openstack.org/pipermail/openstack-discuss/2019-November/010938.html

Change-Id: Ie57c05ff39492f99d66c05cb59e32546a4aa8322