translation job: Fix horizon installation
Horizon was added to upper-constraints.txt in stable branches
recently. This causes a failure in horizon installation in
the translation jobs like [1]. The script tries to install horizon
from the git repo but it conflicts with upper-constraints.

This commit updates the horizon installation to what tox does,
i.e. to install requirements with upper-constraints first and
then install horizon from the git repo.

Note that python modules from test-requirements.txt are now installed.
django-babel (or enmerkar in near future) is what we need here and
it is not a horizon runtime requirement so including test-requirements.txt
allows the horizon team to move django-babel to test-requirements.txt
in future. [2]

[1] https://zuul.opendev.org/t/openstack/build/9d4055098385422580f031c0ee668d2b
[2] http://lists.openstack.org/pipermail/openstack-discuss/2019-December/011783.html

Change-Id: Ibdb63d83b368709fdb1a69da3ce4c60d908c0edb