Add CentOS 8 openafs package build job
Split the extant job into separate centos7 and centos8 parts (this is
so we can easily just rm the centos7 path when it's not required,
rather than making some mashed up combined job), and add a centos8
job to build openafs packages for centos-8.

This is in preparation for making centos-8 wheels; we copy the wheel
directly from the building host so it needs an openafs client.

Change-Id: I2d6392879aac85431253284f217b9daa4a9af94d