Use python3 in release-notes-jobs
release-notes-jobs project-template used in stable branches like rocky
uses python2, but it refers to the master branch of the requirements repo
which now supports only python3. As a result build-openstack-releasenotes
job fail. The releasenotes build is always run against the master branch,
so we can use python3 in release-notes-jobs project-template.

release-notes-jobs and release-notes-jobs-python3 are now similar,
but it would be good to keep both as release-notes-jobs is referred
in stable branches in many repositories.

Change-Id: I708b20424a7b044cac3ea589714235fce5fae546