Also trigger Github mirroring on tags
We transitioned from Gerrit-wide replication to using repo-specific
jobs for Github mirroring of official OpenStack repositories.

The job was only triggered by merges ('post' pipeline), which means
that new tags are not replicated until the next merge.

Since that can take a while for less-active projects, we should
also trigger that job when a tag is pushed ('pre-release' and
'release' pipelines).

Change-Id: Icfb5d4053b59074c5fc5a366557b71aec4039c1d