Fix networking bgpvpn tempest plugin to installed in tempest venv
networking bgpvpn tempest plugin is in-tree plugin till stein and
used to be run with all-plugins tox env which does not work now for
python2.7 jobs. Temepst venv is installed with py3 and all-plugins
no longer detect the system wide installed tempest plugins.

This commit makes to installed networking bgpvpn as tempest plugins so
that Tempest venv can detect it and run tests with 'all' tox env.

Partial-Bug: #1880217

Change-Id: I2a138b859798c63519257fa5986a3bdcf43ded2c