Move wheel build jobs out of project-config
Move the wheel build jobs out of project config so we can run
speculative testing (project-config is a protected repo).
requirements project has these jobs in it's gate.

The only change here is they are renamed to "wheel-cache" from
"wheel-mirror"; both to not conflict but also I think cache describes
what it does a bit better -- we put the cache on the mirror but we're
not mirroring something created externally.

See-Also: https://review.opendev.org/731401

Change-Id: Ib41607c4842dccf27b711947bf356ef5d74a1059