Allow other projects to run wheel cache jobs
Remove the job restriction so we can run these jobs in this gate when
we update the roles/playbooks.  The restriction is moved to the
publish job with the dependent change.

Add dependency on requirements so it is available outside that
repository.

The system-config roles aren't required.  I think this was from when
we used to directly copy to afs from the build nodes and required the
openafs-client install roles.  Now the executor copies the results via
the publish jobs.

Depends-On: https://review.opendev.org/731632

Change-Id: Ica59634eec533db20fe8b1dd148ee9b4d041fd3c