openafs-rpm-package-build: new package build roles
The extant roles are not really right.  They run in periodic and poll
tarballs to see if they should build anything.  They are not gate
tested.  I do not know why I did it like this; but it was a long time
ago.

We should have roles that build in the check gate, and upload only
when the client version changes.

Additionally, I found that the upstream .srpm we were building from is
not quite right; it won't build on ARM64 for example due to the way
it's autoconf machinery has been generated.  The best way is to build
your own srpm from the upstream source using the method linked in this
change.

This builds and publishes artifacts for centos7 x86 and centos8
arm64/x86.  A follow-on will publish these.

Change-Id: Iefb01ba48a457388bb5412ddad3d7dafb00d620c